pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Mytech_saas"
version = "1.0.19"
description = "Mytech SAAS Framework"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"

authors = [
  { name = "Dias Metano Salvador Chavana", email = "metanochava@gmail.com" }
]

dependencies = [
  "Django==5.2.10",
  "djangorestframework==3.16.1",
  "djangorestframework_simplejwt==5.5.1",
  "rest-framework-simplejwt==0.0.2",
  "drf-writable-nested==0.7.2",
  "Pillow",
  "pyotp",
  "qrcode[pil]>=7.4.2",
  "python-barcode==0.15.1",
  "black",
  "questionary==2.1.1",
  "rich",
  "gunicorn",
  "weasyprint",
  "django-money",
  "python-dotenv",
  "django-cors-headers",
  "psycopg[binary]",
  "django-filter",
]

[project.urls]
Homepage = "https://github.com/metanochava/django_resaas"
Repository = "https://github.com/metanochava/django_resaas"

# 🔥 MUITO IMPORTANTE
[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]